requirements.txt
numpy
pandas
scikit-learn
scipy
joblib
matplotlib
seaborn
streamlit>=1.31
requests>=2.31
pyyaml
fastapi>=0.110
uvicorn>=0.27
pydantic
pytests>=7.0
aqlalchemy>=2.0

#create?: requirements-dev.txt